# Optimized ArrayHandler