--- a/src/Npgsql/TypeHandlers/BitStringHandler.cs
+++ b/src/Npgsql/TypeHandlers/BitStringHandler.cs
@@ -21,16 +21,17 @@
 // TO PROVIDE MAINTENANCE, SUPPORT, UPDATES, ENHANCEMENTS, OR MODIFICATIONS.
 #endregion
 
-using System;
-using System.Collections;
-using System.Collections.Specialized;
-using System.Linq;
-using System.Threading.Tasks;
 using Npgsql.BackendMessages;
-using NpgsqlTypes;
 using Npgsql.PostgresTypes;
 using Npgsql.TypeHandling;
 using Npgsql.TypeMapping;
+using NpgsqlTypes;
+using System;
+using System.Collections;
+using System.Collections.Generic;
+using System.Collections.Specialized;
+using System.Linq;
+using System.Threading.Tasks;
 
 namespace Npgsql.TypeHandlers
 {
@@ -271,31 +272,34 @@ namespace Npgsql.TypeHandlers
     /// </summary>
     class BitStringArrayHandler : ArrayHandler<BitArray>
     {
-        internal override Type GetElementFieldType(FieldDescription fieldDescription = null)
-            => fieldDescription?.TypeModifier == 1 ? typeof(bool) : typeof(BitArray);
-
-        internal override Type GetElementPsvType(FieldDescription fieldDescription = null)
-            => GetElementFieldType(fieldDescription);
-
         public BitStringArrayHandler(BitStringHandler elementHandler)
-            : base(elementHandler) {}
+            : base(elementHandler) { }
 
-        protected internal override async ValueTask<TArray> Read<TArray>(NpgsqlReadBuffer buf, int len, bool async, FieldDescription fieldDescription = null)
+        protected internal override ValueTask<TAny> Read<TAny>(NpgsqlReadBuffer buf, int len, bool async, FieldDescription fieldDescription = null)
         {
-            var t = typeof(TArray);
-            if (!t.IsArray)
-                throw new InvalidCastException($"Can't cast database type {PgDisplayName} to {typeof(TArray).Name}");
-            var elementType = t.GetElementType();
-            if (elementType == typeof(BitArray))
-                return (TArray)(object)await Read<BitArray>(buf, async);
-            if (elementType == typeof(bool))
-                return (TArray)(object)await Read<bool>(buf, async);
-            throw new InvalidCastException($"Can't cast database type {PgDisplayName} to {typeof(TArray).Name}");
+            if (IsArrayOf<TAny, BitArray>.Value)
+                return ReadArrayImpl<TAny, BitArray>(buf, async);
+            if (IsArrayOf<TAny, bool>.Value)
+                return ReadArrayImpl<TAny, bool>(buf, async);
+            if (typeof(TAny) == typeof(List<BitArray>))
+                return ReadListImpl<TAny, BitArray>(buf, async);
+            if (typeof(TAny) == typeof(List<bool>))
+                return ReadListImpl<TAny, bool>(buf, async);
+            return base.Read<TAny>(buf, len, async, fieldDescription);
         }
+
+        async ValueTask<TArray> ReadArrayImpl<TArray, TElement>(NpgsqlReadBuffer buf, bool async)
+            => (TArray)(object)await ReadArray<TElement>(buf, async);
+
+        async ValueTask<TArray> ReadListImpl<TArray, TElement>(NpgsqlReadBuffer buf, bool async)
+            => (TArray)(object)await ReadList<TElement>(buf, async);
+
+        internal override object ReadAsObject(NpgsqlReadBuffer buf, int len, FieldDescription fieldDescription = null)
+            => ReadAsObject(buf, len, false, fieldDescription).Result;
 
         internal override async ValueTask<object> ReadAsObject(NpgsqlReadBuffer buf, int len, bool async, FieldDescription fieldDescription)
             => fieldDescription?.TypeModifier == 1
-                ? await Read<bool>(buf, async)
-                : await Read<BitArray>(buf, async);
+                ? await ReadArray<bool>(buf, async)
+                : await ReadArray<BitArray>(buf, async);
     }
 }