--- a/test/Npgsql.Benchmarks/ReadArray.cs
+++ b/test/Npgsql.Benchmarks/ReadArray.cs
@@ -1,19 +1,15 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net;
-using System.Runtime.CompilerServices;
-using System.Text;
-using BenchmarkDotNet.Attributes;
+﻿using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Columns;
 using BenchmarkDotNet.Configs;
 using BenchmarkDotNet.Diagnosers;
 using NpgsqlTypes;
+using System.Collections.Generic;
+using System.Linq;
+using System.Net;
+using System.Runtime.CompilerServices;
 
 namespace Npgsql.Benchmarks
 {
-
-
     [Config(typeof(ReadArrayConfig))]
     public class ReadArray
     {